+++ src/data/autocomplete-value-sets.ts
@@ -135,3 +135,4 @@ export const RarityValueSet: IValueSet = {
 export const BoolValueSet: IValueSet = {name: 'bool', values: [{name: 'true'}, {name: 'false'}]};
 
 export const TagsValueSet: IValueSet = {name: 'tags-list', values: [{name: 'loading...'}]};
+export const StoresValueSet: IValueSet = {name: 'stores-list', values: [{name: 'loading...'}]};