+++ src/data/xml-data/store-elements.ts
@@ -1,0 +1,7 @@
+import {XmlTag} from './helpers';
+
+export const storeTags: XmlTag[] = [
+  {name: 'customStore', attributes: [{name: 'id'}]},
+  {name: 'resource', attributes: []},
+  {name: 'category', attributes: [{name: 'type'}]},
+];