+++ src/models/ftl-custom-store.ts
@@ -1,0 +1,5 @@
+import {FtlValue} from './ftl-value';
+
+export class FtlCustomStore extends FtlValue {
+  readonly kind = 'Store';
+}